requirements-minimal.txt
# Core web framework
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
sse-starlette

# Data handling
pandas>=2.0.3
pydantic>=2.4.2
requests
python-dotenv
numpy
beautifulsoup4

# Templates and static files
jinja2

# Database
duckdb

# Configuration
pyyaml

# Basic utilities
tqdm

# LLM
openai
tiktoken

# Other dependencies
multiformats
psutil

# Search and NLP
elasticsearch
nltk
rank_bm25

# Additional FastAPI middleware
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Document processing
PyPDF2
python-docx
doc2docx
python-magic